f3probe: rename real size to usable size
f3probe does not try to find the real amount of memory that
a fake card has, but the usable about of memory, that is,
the memory from the first block (i.e. block zero) to
the block before the first failed block.

A fake card may have more memory, but it would be spread among
failed blocks, so it is not usable.