f3probe: better ballance writes and resets while bisecting
f3probe was using just 3 writes to make a decision on how to
balance resets and writes.
When the first 3 writes were faster than normal,
the next pass would go for an unrealistic number of writes.

This patch requires at least 10 writes;
in practice, it will uses 15 writes.
The larger number of writes allows for a better balance.